[idea] #337 Refactored Settings class, used builder pattern
fixes #337
